chore: rename files to inspectorMode instead of fieldTagging
