+++ src/__tests__/init.spec.ts
@@ -1,8 +1,8 @@
 // @vitest-environment jsdom
 import { describe, it, expect } from 'vitest';
 
-import { FieldTagging } from '../fieldTagging';
 import { ContentfulLivePreview } from '../index';
+import { FieldTagging } from '../inspectorMode';
 
 describe('init', () => {
   it('returns a Promise that resolves to a LivePreview instance when running in a browser environment', async () => {
