+++ src/inspectorMode.ts
